Fixes
Added command Tell
Added serializers
Added commands config
Update Commands API

+++ synapse-api/src/main/java/sawfowl/synapse/api/commands/settings/CommandSettings.java
@@ -56,9 +56,9 @@ public interface CommandSettings {
 
 		Builder setPrice(CommandPrice price);
 
-		Builder setDelay(int delay);
+		Builder setDelay(long delay);
 
-		Builder setCooldown(int cooldown);
+		Builder setCooldown(long cooldown);
 
 		Builder setIgnoreDelay(String permission);
